% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pyth.R
\name{Pythagoras}
\alias{Pythagoras}
\title{================================PYTHAGORAS FUNCTION===============================
DISCRIPTION:
The square of the hypotenuse (the side opposite the right angle of a
trianble, c) is equal to the sum of the squares of the other two sides (a & b).}
\usage{
Pythagoras(a, b)
}
\arguments{
\item{a}{& b}
}
\value{
Length of c
}
\description{
================================PYTHAGORAS FUNCTION===============================
DISCRIPTION:
The square of the hypotenuse (the side opposite the right angle of a
trianble, c) is equal to the sum of the squares of the other two sides (a & b).
}

